別ファイルにしていたが一本化した
